SD2 Move world_map_kalimdor to separate file
